pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "otfm"
version = "0.1.0"
description = "Optimal Transport couplings for Flow Matching on 2D toy data."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
    { name = "Giovanni Dettori" },
    { name = "Mael Tremouille" },
]
dependencies = [
    "numpy>=1.26",
    "scipy>=1.11",
    "matplotlib>=3.8",
    "pandas>=2.1",
    "scikit-learn>=1.3",
    "jax>=0.4.28",
    "optax>=0.2.2",
    "ott-jax>=0.4.8",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "ruff>=0.4",
    "black>=24.3",
    "nbstripout>=0.7",
]
notebook = [
    "jupyter>=1.0.0",
    "ipykernel>=6.29",
]

[tool.setuptools.packages.find]
include = ["otfm*"]
exclude = ["tests*", "experiments*", "scripts*", "report*", "slides*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]
ignore = ["E501"]  # line length handled by formatter

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-q --strict-markers"
